+++ src/Components/AppBar.js
@@ -9,6 +9,7 @@ import Box from '@mui/material/Box';
 import Brightness4Icon from '@mui/icons-material/Brightness4';
 import Brightness7Icon from '@mui/icons-material/Brightness7';
 import { useTheme } from '@mui/material/styles';
+import { ColorModeContext } from './ColorMode';
 
 import { Translate } from './Translate';
 
@@ -39,7 +40,7 @@ export const AppBar = ({ handleDrawerToggle, isLargerScreen }) => {
           </Typography>
           <IconButton
             sx={{ ml: 1 }}
-            // onClick={colorMode.toggleColorMode}
+            onClick={colorMode.toggleColorMode}
             color="inherit"
           >
             {theme.palette.mode === 'dark' ? (
