+++ src/Components/ColorMode.js
@@ -1,0 +1,38 @@
+import React from 'react';
+
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import useMediaQuery from '@mui/material/useMediaQuery';
+import CssBaseline from '@mui/material/CssBaseline';
+
+export const ColorModeContext = React.createContext({
+  toggleColorMode: () => {},
+});
+
+export const ColorMode = ({ children }) => {
+  const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
+  const preferredMode = prefersDarkMode ? 'dark' : 'light';
+  const [mode, setMode] = React.useState(preferredMode);
+  const theme = React.useMemo(() => createTheme({ palette: { mode } }), [mode]);
+
+  React.useEffect(() => {
+    setMode(preferredMode);
+  }, [preferredMode]);
+
+  const colorMode = React.useMemo(
+    () => ({
+      toggleColorMode: () => {
+        setMode(prevMode => (prevMode === 'light' ? 'dark' : 'light'));
+      },
+    }),
+    [],
+  );
+
+  return (
+    <ColorModeContext.Provider value={colorMode}>
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        {children}
+      </ThemeProvider>
+    </ColorModeContext.Provider>
+  );
+};
